build: bundle assets into dist, load from there
Raw data about the liturgical calendar and the feasts of the saints is
stored in `/assets`. `loadSaints()` tries to load from there by giving a
relative path. When using this library as a dependency (or in any process
not running at this repository's root) that relative path lookup will
fail.

Here, we:
- load the files from absolute paths based on the `__dirname` of the
  source file, and
- add an asset copying step to the build script so that `/assets` ends
  up in the `/dist`.

For the former, my impression is that Node handles the
cross-platform-ing of the paths for you, so we can naively concatenate.
(Many sources suggest using the `path` module, but given that the
hard-coded POSIX-style paths were working fine, this may not be strictly
necessary?)

For the latter, we are forced to add a dev dependency (`cpx`) for
copying the files over. Writing a simple `cp` or `rsync` would work on
POSIX systems, but not Windows.