calendar: crude i18n for dynamic dates
Functionally does what f0acebd did for the ones with fixed dates. We
don't load the internationalized titles from files here, instead have an
extra `title_` attribute for `nl_NL` names. If the language is set to
that, and we have a `title_` field, replace the `title`'s contents with
that.

This should probably be restructured, either to have a `titles` object
with `lang` keys, or to put these events into the translation files also
and look them up there.

(But we consider that restructuring out of scope for now.)