build: revert to loading assets from /assets
This is a partial revert of 4fdee83. We shouldn't need to move the
assets folder into dist per se. As long as we point to it correctly
(still with `__dirname` from the prior commit) it should load fine, as
the assets folder at the root level is already included in the files for
the package as specified in `package.json`.