calendar: use title from specified language (#2)
* calendar: use title from specified language

Previously, we always used the Latin `id` in the `title` field. Here, we
respect the `lang` argument and use it to load a translation file if
needed.

Replacing of the space after "H." (Dutch for "S." saint prefix) seemed
vestigial, so we take it out.

* calendar: crude i18n for dynamic dates

Functionally does what f0acebd did for the ones with fixed dates. We
don't load the internationalized titles from files here, instead have an
extra `title_` attribute for `nl_NL` names. If the language is set to
that, and we have a `title_` field, replace the `title`'s contents with
that.