build: bundle assets into dist, load from there (#1)
* build: bundle assets into dist, load from there

Raw data about the liturgical calendar and the feasts of the saints is
stored in `/assets`. `loadSaints()` tries to load from there by giving a
relative path. When using this library as a dependency (or in any process
not running at this repository's root) that relative path lookup will
fail.

Here, we load the files from absolute paths based on the `__dirname` of the 
source file.