dist: ensure ts typedefs get shipped
Typescript type definitions (that already get generated during tsc
compilation) were not being included in the package. We make two changes
to resolve this:

First, we list `*.d.ts` files for inclusion in the package. The build
process already generates these.

...Except for `src/types.d.ts`, which the compiler ignores because of
its `.d.ts` extension (as opposed to plain `.ts`). Other type definition
files still try referring to it (as `./types`) but that lookup fails
because no `dist/types.*` file got created.
(Arguably it's a `tsc` bug that it creates files with broken imports.)

So, secondly, we rename `types.d.ts` to `types.ts`, causing it to be
picked up for processing by `tsc` and resulting in `dist/types.*` files.