dist: include .d.ts files from src/
Reverts the extension renaming done in 85a3d6b. Instead includes
`.d.ts` files from `src/` in the package files and updates the imports
referring to the `src/types.d.ts` to use a relative path instead.

(This works now because `src/` and `dist/` will contain files at the
same depth (& in the same structure), but may fail to resolve if that
ever changes.)